requirements.txt
asgiref
certifi
charset-normalizer
colorama
contourpy
cycler
Django
experimental
filelock
fonttools
fsspec
gitdb
GitPython
idna
Jinja2
kiwisolver
MarkupSafe
matplotlib
mpmath
networkx
numpy
opencv-python
packaging
pandas
pillow
psutil
py-cpuinfo
pyparsing
python-dateutil
pytz
PyYAML
requests
scipy
seaborn
six
smmap
sqlparse
sympy
thop
torch
torchvision
tqdm
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
